Display a message if no sequence is lodaded. Create a component for IconMessages
